+++ packages/expo-router/src/getServerManifest.ts
@@ -37,10 +37,6 @@ export interface RouteRegex {
   re: RegExp;
 }
 
-function isApiRoute(route: RouteNode) {
-  return !route.children.length && !!route.contextKey.match(/\+api\.[jt]sx?$/);
-}
-
 function isNotFoundRoute(route: RouteNode) {
   return route.dynamic && route.dynamic[route.dynamic.length - 1].notFound;
 }
@@ -64,17 +60,31 @@ export function getServerManifest(route: RouteNode): ExpoRouterServerManifestV1 
       return route.children.map((child) => getFlatNodes(child)).flat();
     }
 
-    const key = getContextKey(route.contextKey).replace(/\/index$/, '') ?? '/';
+    // API Routes are handled differently to HTML routes because they have no nested behavior.
+    // An HTML route can be different based on parent segments due to layout routes, therefore multiple
+    // copies should be rendered. However, an API route is always the same regardless of parent segments.
+    let key: string;
+    if (route.type === 'api') {
+      key = getContextKey(route.contextKey).replace(/\/index$/, '') ?? '/';
+    } else {
+      key = getContextKey(route.route).replace(/\/index$/, '') ?? '/';
+    }
     return [[key, route]];
   }
 
   // Remove duplicates from the runtime manifest which expands array syntax.
-  const flat = uniqueBy(getFlatNodes(route), ([key]) => key)
+  const flat = getFlatNodes(route)
     .sort(([, a], [, b]) => sortRoutes(b, a))
     .reverse();
 
-  const apiRoutes = flat.filter(([, route]) => isApiRoute(route));
-  const otherRoutes = flat.filter(([, route]) => !isApiRoute(route));
+  const apiRoutes = uniqueBy(
+    flat.filter(([, route]) => route.type === 'api'),
+    ([path]) => path
+  );
+  const otherRoutes = uniqueBy(
+    flat.filter(([, route]) => route.type === 'route'),
+    ([path]) => path
+  );
   const standardRoutes = otherRoutes.filter(([, route]) => !isNotFoundRoute(route));
   const notFoundRoutes = otherRoutes.filter(([, route]) => isNotFoundRoute(route));
 
@@ -97,6 +107,7 @@ function getMatchableManifestForPaths(
   return paths.map((normalizedRoutePath) => {
     const matcher: ExpoRouterServerManifestV1Route = getNamedRouteRegex(
       normalizedRoutePath[0],
+      getContextKey(normalizedRoutePath[1].route),
       normalizedRoutePath[1].contextKey
     );
     if (normalizedRoutePath[1].generated) {
@@ -106,14 +117,15 @@ function getMatchableManifestForPaths(
   });
 }
 
-export function getNamedRouteRegex(
+function getNamedRouteRegex(
   normalizedRoute: string,
-  page: string
+  page: string,
+  file: string
 ): ExpoRouterServerManifestV1Route {
   const result = getNamedParametrizedRoute(normalizedRoute);
   return {
-    file: page,
-    page: page.replace(/\.[jt]sx?$/, ''),
+    file,
+    page,
     namedRegex: `^${result.namedParameterizedRoute}(?:/)?$`,
     routeKeys: result.routeKeys,
   };